vhdl: fixing counter incrementation
Rebuilding the examples, I had issues at incrementation of a vector type.
Actually counting and carry over did not work auto-magically for me. First I
had an error message, bringing it to work it did not count as expected.

Another item for me was the check "(COUNT_MOORE_REG +1) = M - 1", does this
actually make sense, or do I missunderstand it, in my opinion it should be
either sustraction or addition by one. Performing both neutralizes itself, or
am I wrong?

The patch should is my approach to fix this issue. Please, give it a chance and
verify if the patch works out. I would really appreciate contribution to your
project.